+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Head
-* Removed: `jsesc` devDependency.
+
+* Added: `rule-no-shorthand-property-overrides` rule.
 * Added: `ruleTester` to `stylelint.utils` for use by authors of custom rules.
+* Removed: `jsesc` devDependency.
 
 # 0.3.2
+
 * Fixed: `hierarchicalSelectors` bug in `indentation` rule.
 
 # 0.3.1